untuk fadil
fadil jancok udah tak bilang pake longblob

+++ admin/modul/mod_galeri/hapus_galeri.php
@@ -3,12 +3,10 @@ $id = $_GET['id'];
 
 $sql = mysqli_query($db, "SELECT * FROM galeri WHERE id_galeri='$id'");
 
-$r = mysqli_fetch_array($sql);
-
-unlink("././img_galeri/".$r['nama_foto']);
-
-mysqli_query($db, "DELETE FROM galeri WHERE id_galeri='$id'");
-
-echo "<script>alert('Data galeri Berhasil dihapus'); window.location = 'dashboard.php?hal=galeri'</script>";
-
+if(mysqli_num_rows($sql) > 0){
+    mysqli_query($db, "DELETE FROM galeri WHERE id_galeri='$id'");
+    echo "<script>alert('Data galeri berhasil dihapus'); window.location = 'dashboard.php?hal=galeri'</script>";
+} else {
+    echo "<script>alert('Data tidak ditemukan'); window.location = 'dashboard.php?hal=galeri'</script>";
+}
 ?>
